More clippy

+++ singer-utils/src/structs.rs
@@ -40,4 +40,4 @@ pub struct ChipChallenges {
 pub type UInt64 = UInt<64, VALUE_BIT_WIDTH>;
 pub type PCUInt = UInt64;
 pub type TSUInt = UInt<48, 48>;
-pub type StackUInt = UInt<{ EVM_STACK_BIT_WIDTH as usize }, { VALUE_BIT_WIDTH as usize }>;
+pub type StackUInt = UInt<{ EVM_STACK_BIT_WIDTH }, { VALUE_BIT_WIDTH }>;
